% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_stain_d}
\alias{scale_colour_stain_d}
\title{Discrete colour stain}
\usage{
scale_colour_stain_d(primary = "lemon", other = "wood", direction = 1, ...)
}
\arguments{
\item{primary}{The default primary color name for the stain palette is "lemon"}

\item{other}{The default primary color name for the stain palette is "wood"}

\item{direction}{Sets the order of the colours in the scale. If 1, it gives the default order; if less than 0 the order of colours is reversed}

\item{...}{Other parameters that support the scale_colour_discrete function}
}
\value{
It will return corresponding hexcode of the colour name to the ggplot
}
\description{
Discrete colour stain
}
\examples{
scale_colour_stain_d(primary = "lemon", other = "wood", direction = 1)
scale_colour_stain_d(primary = "lemon", other = "pink", direction = 1)
}
