% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_stain_seq_c}
\alias{scale_colour_stain_seq_c}
\title{Continuous scale for Stain}
\usage{
scale_colour_stain_seq_c(name = "teal", direction = 1, amount = 0.6, ...)
}
\arguments{
\item{name}{The name parameter sets the color for the plot. This will only allow to use colors from the palette, else throws an error.}

\item{direction}{The default direction parameter produce the high luminance colour for the top of the scale.(i.e. lighter colours correspond to higher values).}

\item{amount}{this argument to the function allows the user to define the amount of lightening}

\item{...}{Other parameters that support the scale_colour_continous function}
}
\value{
It will return corresponding hexcode of the colour name to the ggplot
}
\description{
Continuous scale for Stain
}
\examples{
scale_colour_stain_seq_c(name = "lemon", direction = 1, amount = 0.7)
scale_colour_stain_seq_c(name = "skin", direction = 1, amount = 0.7)
}
