% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_fill_stain_d}
\alias{scale_fill_stain_d}
\title{#' Discrete fill stain}
\usage{
scale_fill_stain_d(primary = "lemon", other = "wood", direction = 1, ...)
}
\arguments{
\item{primary}{The default primary color name for the stain palette is "lemon"}

\item{other}{The default primary color name for the stain palette is "wood"}

\item{direction}{Sets the order of the colours in the scale. If 1, the default; if less than 0 the order of colours is reversed}

\item{...}{Other parameters that support the scale_fill_discrete function}
}
\value{
It will return corresponding hexcode of the colour name to the ggplot
}
\description{
#' Discrete fill stain
}
\examples{
 scale_fill_stain_d(primary = "lemon", other = "wood", direction = 1)
 scale_fill_stain_d(primary = "lemon", other = "pink", direction = 1)
}
