% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{stain_colours}
\alias{stain_colours}
\title{Stain Colours}
\format{
An object of class \code{list} of length 6.
}
\usage{
stain_colours
}
\description{
Stain Colours
}
\examples{
stain_colours
}
\keyword{datasets}
