% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{stain_palette}
\alias{stain_palette}
\title{stain palette}
\usage{
stain_palette(primary = "lemon", other = "wood", direction = 1)
}
\arguments{
\item{primary}{The default primary color of the stain palette is lemon(Yellow)}

\item{other}{The Other color of the stain palette is wood(dark Brown)}

\item{direction}{Sets the order of the colours in the scale. If 1, it gives the default order; if less than 0 the order of colours is reverse}
}
\value{
It will return corresponding hexcode of the colour name to the ggplot
}
\description{
stain palette
}
\examples{
stain_palette(primary = "lemon", other = "wood", direction = 1)(2)
stain_palette(primary = "lemon", other = "pink", direction = 1)(2)
}
