% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_stain_seq_c}
\alias{scale_colour_stain_seq_c}
\title{Continuous scale for Stain}
\usage{
scale_colour_stain_seq_c(name = "teal", direction = 1, amount = 0.6, ...)
}
\arguments{
\item{name}{The default primary color name for the stain palette is "lemon"}

\item{direction}{The default primary color name for the stain palette is "wood"}

\item{amount}{Sets the order of the colours in the scale. If 1, the default; if less than 0 the order of colours is reversed}

\item{...}{Other parameters that support the scale_colour_continous function}
}
\value{
It will return corresponding hexcode of the colour name to the ggplot
}
\description{
Continuous scale for Stain
}
\examples{
scale_colour_stain_seq_c(name = "lemon", direction = 1, amount = 0.7)
scale_colour_stain_seq_c(name = "skin", direction = 1, amount = 0.7)
}
